refactor: Create `useSchema` hook to make validation logic and `InputField` components shareable (#3526)
