+++ editor.planx.uk/src/@planx/components/List/schemas/ResidentialUnits/GLA/New.ts
@@ -1,4 +1,4 @@
-import { Schema } from "@planx/components/List/model";
+import { Schema } from "@planx/components/shared/Schema/model";
 
 export const ResidentialUnitsGLANew: Schema = {
   type: "New residential unit type",
