+++ editor.planx.uk/src/@planx/components/List/schemas/mocks/MaxOne.ts
@@ -1,4 +1,4 @@
-import { Schema } from "@planx/components/List/model";
+import { Schema } from "@planx/components/shared/Schema/model";
 
 import { Props } from "../../Public";
